Fixing various type warnings. We don't today have these warnings enabled in our code but it'd be nice to remove them from the list at some point. This starts us down that path by hopefully making it easier for code authors to copy/paste or at least notice the right patterns.

+++ iree/base/init.cc
@@ -28,7 +28,7 @@ void InitializeEnvironment(int* argc, char*** argv) {
     auto positional_args = absl::ParseCommandLine(*argc, *argv);
     if (positional_args.size() < *argc) {
       // Edit the passed argument refs to only include positional args.
-      *argc = positional_args.size();
+      *argc = static_cast<int>(positional_args.size());
       for (int i = 0; i < *argc; ++i) {
         (*argv)[i] = positional_args[i];
       }
